pyproject.toml
[project]
name = "bank-statement-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "jinja2>=3.1.6",
    "openpyxl==3.1.2",
    "pandas==2.1.3",
    "pdf2image>=1.17.0",
    "pdfplumber==0.10.2",
    "pydantic==2.4.2",
    "pydantic-settings>=2.2.1",
    "pymupdf>=1.27.2.3",
    "pytesseract>=0.3.13",
    "python-dateutil==2.8.2",
    "python-multipart==0.0.18",
    "uvicorn==0.24.0.post1",
]
